Harden the video upload APIs (#509)
Widen the multipart size and count parameters to long. The AT Protocol
integer type is a signed 64 bit value, so a partSizeBytes or partCount
over int.MaxValue threw a JsonException rather than deserializing.

UploadVideo() no longer throws when a 409 already_exists response carries
a jobId which is not a usable string. The value comes from the service's
error extension data and was read with GetString()!, so a number, null, an
array, an object or a blank string threw rather than returning the conflict
to the caller.

UploadStatus.ReceivedParts is now a defensive copy. It previously exposed
the caller's or the deserializer's own collection through an
IReadOnlyCollection<int>, which could be cast back to List<int> and mutated.

UploadVideo() now validates mimeType rather than letting an invalid value
surface as an undocumented FormatException, and StartUpload() and
UploadVideo() now measure mimeType and name in UTF-8 bytes, as the lexicon
does, against named Maximum constants.

Correct the Unknown state semantics. The lexicon specifies that any state
which is not a known value indicates the job is still in process, so the
docs, both samples and the documentation no longer treat it as a failure.
Add UploadStatus.RawState and AbortUploadResponse.RawState so an unmapped
state can still be read, as JobStatus.RawState already allowed.

+++ src/idunno.Bluesky/Video/Agent/UploadPart.cs
@@ -27,7 +27,7 @@ public partial class BlueskyAgent
     [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Catch all exception logging")]
     public async Task<AtProtoHttpResult<UploadPartResponse>> UploadPart(
         string jobId,
-        int part,
+        long part,
         byte[] bytes,
         TimeSpan? timeout = null,
         CancellationToken cancellationToken = default)
